Changed type for finance: Double -> BigDecimal

+++ app/src/main/java/yandexschool/dmpolyakov/money/ui/tracker/TrackerPresenter.kt
@@ -5,6 +5,7 @@ import yandexschool.dmpolyakov.money.BALANCE_IN_RUBBLES
 import yandexschool.dmpolyakov.money.Currency
 import yandexschool.dmpolyakov.money.navigation.MainRouter
 import yandexschool.dmpolyakov.money.ui.base.mvp.BaseMvpPresenter
+import java.math.BigDecimal
 import javax.inject.Inject
 
 
@@ -24,7 +25,7 @@ class TrackerPresenter @Inject constructor(router: MainRouter) : BaseMvpPresente
         viewState?.showBalance(balance, Currency.Rubble)
     }
 
-    private fun getBalanceInRubble(): Double {
+    private fun getBalanceInRubble(): BigDecimal {
         return BALANCE_IN_RUBBLES
     }
