+++ app/src/main/java/yandexschool/dmpolyakov/money/ui/tracker/TrackerView.kt
@@ -2,8 +2,9 @@ package yandexschool.dmpolyakov.money.ui.tracker
 
 import yandexschool.dmpolyakov.money.Currency
 import yandexschool.dmpolyakov.money.ui.base.mvp.BaseMvpView
+import java.math.BigDecimal
 
 interface TrackerView : BaseMvpView {
 
-    fun showBalance(count: Double, currency: Currency = Currency.Rubble)
+    fun showBalance(count: BigDecimal, currency: Currency = Currency.Rubble)
 }
